refactor: import run_coroutine_sync from core.async_bridge in substrate
run_coroutine_sync's real implementation already lives in
polylogue.core.async_bridge (dependency-free: only asyncio/threading);
polylogue.api.sync.bridge is a compatibility re-export for callers that
already depend on the Polylogue facade being available. Four substrate
call sites (storage/embeddings/materialization.py,
insights/correlation_view.py, insights/registry.py,
pipeline/ingest_support.py) were still importing it via
polylogue.api.sync.bridge, forcing polylogue.api.__init__ (and the
whole Polylogue async facade) to import even though they never use the
facade. Import the dependency-free module directly instead.

Ref polylogue-exb
