+++ polylogue/storage/embeddings/materialization.py
@@ -1287,7 +1287,7 @@ def embed_session_sync(
     ``fetch_title=True`` issues an extra ``view`` lookup so callers can
     display a friendly label; when False the title field is left ``None``.
     """
-    from polylogue.api.sync.bridge import run_coroutine_sync
+    from polylogue.core.async_bridge import run_coroutine_sync
 
     title: str | None = None
     if fetch_title:
